# nexora-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexora OS API",
    "version": "1.0.0",
    "description": "Nexora OS microservices. Served via API gateway at http://localhost:3000. Bearer token (Keycloak realm `nexora`)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local gateway"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/tenants": {
      "get": {
        "summary": "GET /api/v1/tenants",
        "operationId": "get__api_v1_tenants",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/admin/tenants": {
      "post": {
        "summary": "POST /api/v1/admin/tenants",
        "operationId": "post__api_v1_admin_tenants",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}